fix: Security upgrade jsonwebtoken to 9.0.0 (#8431)
