+++ package.json
@@ -37,7 +37,7 @@
     "graphql-tag": "2.12.6",
     "graphql-relay": "0.10.0",
     "intersect": "1.0.1",
-    "jsonwebtoken": "8.5.1",
+    "jsonwebtoken": "9.0.0",
     "jwks-rsa": "2.1.4",
     "ldapjs": "2.3.3",
     "lodash": "4.17.21",
